tools/dev/requirements.txt
websockets>=12.0,<16.0
PyYAML>=6.0,<7.0
prompt_toolkit>=3.0,<4.0
